select random sketch
